use uint64 instead uint32 for FramMetrics

+++ interfaces/innerkits/wm/wm_common.h
@@ -1557,9 +1557,9 @@ struct ExceptionInfo : public Parcelable {
  */
 struct FrameMetrics : public Parcelable {
     bool firstDrawFrame_ = false;
-    uint32_t inputHandlingDuration_ = 0;
-    uint32_t layoutMeasureDuration_ = 0;
-    uint32_t vsyncTimestamp_ = 0;
+    uint64_t inputHandlingDuration_ = 0;
+    uint64_t layoutMeasureDuration_ = 0;
+    uint64_t vsyncTimestamp_ = 0;
 
     /**
      * @brief Marshalling FrameMetrics.
@@ -1570,9 +1570,9 @@ struct FrameMetrics : public Parcelable {
     bool Marshalling(Parcel& parcel) const override
     {
         return parcel.WriteBool(firstDrawFrame_) &&
-               parcel.WriteUint32(inputHandlingDuration_) &&
-               parcel.WriteUint32(layoutMeasureDuration_) &&
-               parcel.WriteUint32(vsyncTimestamp_);
+               parcel.WriteUint64(inputHandlingDuration_) &&
+               parcel.WriteUint64(layoutMeasureDuration_) &&
+               parcel.WriteUint64(vsyncTimestamp_);
     }
 
     /**
@@ -1585,9 +1585,9 @@ struct FrameMetrics : public Parcelable {
     {
         auto info = new FrameMetrics();
         if (!parcel.ReadBool(info->firstDrawFrame_) ||
-            !parcel.ReadUint32(info->inputHandlingDuration_) ||
-            !parcel.ReadUint32(info->layoutMeasureDuration_) ||
-            !parcel.ReadUint32(info->vsyncTimestamp_)) {
+            !parcel.ReadUint64(info->inputHandlingDuration_) ||
+            !parcel.ReadUint64(info->layoutMeasureDuration_) ||
+            !parcel.ReadUint64(info->vsyncTimestamp_)) {
             delete info;
             return nullptr;
         }
